hf_space_deploy/requirements.txt
gradio==4.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.45.0
peft>=0.12.0
Pillow>=10.0.0
qwen-vl-utils>=0.0.8
accelerate>=0.20.0
datasets>=2.0.0
opencv-python>=4.5.0
